Moved ATRC.h into ATRC\ATRC.h
